Use constructable stylesheets with webpack loader
