+++ components/Accordion/style.css
@@ -1,0 +1,10 @@
+:host {
+  display: flex;
+  flex-direction: column;
+  gap: var(--tcds-space-x-tight);
+}
+
+[part="controls"] {
+  display: flex;
+  justify-content: flex-end;
+}
