+++ components/Accordion/style.light.css
@@ -1,0 +1,7 @@
+tcds-accordion:not(:only-child) {
+  margin: var(--tcds-space-normal) 0;
+}
+
+tcds-accordion:first-child:not(:only-child) {
+  margin: 0 0 var(--tcds-space-normal);
+}
