+++ components/AccordionSection/style.css
@@ -1,0 +1,41 @@
+[part="heading"] {
+  margin: 0;
+}
+
+[part="button"] {
+  appearance: none;
+  background: var(--tcds-color-gray-light);
+  border: 0;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: var(--tcds-size-medium);
+  padding: 0 var(--tcds-space-normal);
+  text-align: left;
+  font-size: var(--tcds-font-size-small);
+  font-family: var(--tcds-font-ui);
+  font-weight: var(--tcds-font-weight-semibold);
+  width: 100%;
+  color: var(--tcds-color-gray-dark);
+}
+
+[part="button"][aria-expanded="true"] {
+  background-color: var(--tcds-color-blue-dark);
+  color: var(--tcds-color-gray-x-light);
+  font-weight: var(--tcds-font-weight-normal);
+}
+
+[part="icon"] {
+  pointer-events: none;
+}
+
+[part="panel"] {
+  overflow: hidden;
+  background-color: var(--tcds-color-gray-x-light);
+  transition: height 0.15s linear;
+}
+
+[part="content"] {
+  padding: var(--tcds-space-normal) var(--tcds-space-loose);
+}
