+++ components/AlertBar/style.light.css
@@ -1,0 +1,44 @@
+/* stylelint-disable property-no-vendor-prefix */
+
+tcds-alert-bar details {
+  display: contents;
+}
+
+tcds-alert-bar details[open] > summary {
+  --tcds-details-icon-rotate: -180deg;
+}
+
+tcds-alert-bar details summary {
+  grid-row: 1 / 2;
+  height: var(--tcds-alert-bar-height);
+  flex-direction: row-reverse;
+  margin-right: var(--tcds-space-normal);
+}
+
+tcds-alert-bar details summary:hover {
+  color: var(--tcds-color-primary);
+}
+
+tcds-alert-bar details[open] summary {
+  box-shadow: inset 0 -4px 0 var(--tcds-color-primary);
+}
+
+tcds-alert-bar details summary::before {
+  -webkit-mask: var(--tcds-icon-chevron-down);
+  mask: var(--tcds-icon-chevron-down);
+  background-color: var(--tcds-color-primary);
+}
+
+tcds-alert-bar details article {
+  grid-row: 2 / -1;
+  grid-column: 1 / -1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  padding: var(--tcds-space-normal) 0;
+  position: sticky;
+  left: 0;
+  max-width: calc(100vw - (var(--site-outer-gutter) * 2));
+  background-color: #fff;
+  border-top: 1px solid var(--tcds-color-gray-light);
+}
