+++ components/Card/style.css
@@ -1,0 +1,78 @@
+:host {
+  display: block;
+  width: 100%;
+  height: 100%;
+}
+
+[part="card"] {
+  display: grid;
+  grid-template-areas: "content";
+  position: relative;
+  background-color: #fff;
+  width: 100%;
+  height: 100%;
+  box-shadow: 0 0 0.625rem 0 rgb(82 82 82 / 16%);
+  border-radius: 10px;
+  overflow: hidden;
+}
+
+:host([orientation="vertical"]:not([no-image])) [part="card"] {
+  grid-template-areas:
+    "image"
+    "content";
+  grid-template-rows: auto 1fr;
+}
+
+:host([orientation="horizontal"]:not([no-image])) [part="card"] {
+  grid-template-columns: var(--tcds-card-image-width, minmax(220px, 1fr)) 3fr;
+  grid-template-rows: 1fr;
+  grid-template-areas: "image content";
+}
+
+:host([size*="small"]) {
+  --tcds-card-image-width: minmax(auto, 220px);
+  --tcds-card-content-padding: var(--tcds-space-loose);
+  --tcds-card-title-font-size: var(--tcds-font-size-medium-plus);
+  --tcds-card-description-font-size: var(--tcds-font-size-small);
+}
+
+:host([size*="large"]) {
+  --tcds-card-title-font-size: var(--tcds-font-size-x-large);
+  --tcds-card-description-font-size: var(--tcds-font-size-medium-plus);
+}
+
+:host([divider]:not([orientation*="horizontal"])) {
+  --tcds-card-content-divider: var(--tcds-space-tight) solid var(--tcds-color-primary);
+}
+
+:host([image*="contain"]) {
+  --tcds-card-image-fit: contain;
+}
+
+:host([image*="half"]) {
+  --tcds-card-image-width: 3fr;
+}
+
+figure {
+  grid-area: image;
+  overflow: hidden;
+}
+
+[part="content"] {
+  grid-area: content;
+  display: flex;
+  flex-direction: column;
+  color: var(--tcds-color-gray);
+  border-top: var(--tcds-card-content-divider, none);
+  padding:
+    var(--tcds-card-content-padding-y, var(--tcds-card-content-padding, var(--tcds-space-x-loose)))
+    var(--tcds-card-content-padding-x, var(--tcds-card-content-padding, var(--tcds-space-x-loose)));
+}
+
+::slotted(img) {
+  transition: transform var(--tcds-animation-expressive-duration) var(--tcds-animation-expressive-easing);
+}
+
+:host(:hover) ::slotted(img) {
+  transform: scale(1.05);
+}
