+++ components/Card/style.light.css
@@ -1,0 +1,71 @@
+tcds-card [slot="tag"] {
+  position: absolute;
+  top: 0;
+  right: var(--tcds-space-loose);
+  background-color: #f3f3f3;
+  text-transform: uppercase;
+  font-family: var(--tcds-font-ui);
+  font-size: var(--tcds-font-size-small);
+  color: var(--tcds-color-gray);
+  border-bottom-right-radius: 8px;
+  border-bottom-left-radius: 8px;
+  padding: var(--tcds-space-x-tight) var(--tcds-space-normal);
+}
+
+tcds-card [slot="image"] {
+  display: block;
+  width: 100%;
+  height: 100%;
+  object-fit: var(--tcds-card-image-fit, cover);
+  object-position: center;
+}
+
+tcds-card [slot="title"] {
+  font-family: var(--tcds-font-copy);
+  font-weight: var(--tcds-font-weight-semibold);
+  font-size: var(--tcds-card-title-font-size, var(--tcds-font-size-large));
+  line-height: var(--tcds-line-height-tight);
+  color: var(--tcds-color-secondary);
+  margin-bottom: var(--tcds-space-tight);
+  text-decoration: none;
+}
+
+tcds-card [slot="title"]:focus {
+  outline: none;
+}
+
+tcds-card [slot="title"][href]::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+}
+
+tcds-card [slot="description"],
+tcds-card [slot="description"] p {
+  font-size: var(--tcds-card-description-font-size, var(--tcds-font-size-medium));
+  line-height: 1.5;
+  color: var(--tcds-color-gray-dark);
+  margin-bottom: var(--tcds-space-normal);
+}
+
+tcds-card [slot="footer"],
+tcds-card::part(footer) {
+  display: flex;
+  gap: var(--tcds-space-normal);
+  margin-top: auto;
+}
+
+tcds-card [slot="footer"] a,
+tcds-card::part(footer) a {
+  text-decoration: none;
+}
+
+tcds-card [slot="description"] a,
+tcds-card [slot="footer"] a {
+  position: relative;
+  margin: -0.5em;
+  padding: 0.5em;
+}
