+++ components/Carousel/style.css
@@ -1,0 +1,98 @@
+:host {
+  display: grid;
+  grid-template-areas:
+    "slides     slides"
+    "play-pause navigation";
+  grid-template-columns: min-content 1fr;
+  align-items: center;
+  gap: var(--tcds-space-loose) 0;
+}
+
+:host([navigation*="top"][navigation*="right"]) {
+  --tcds-carousel-navigation-justify: end;
+
+  grid-template-areas:
+    ".          navigation"
+    "slides     slides"
+    "play-pause .";
+  gap: var(--tcds-space-tight);
+}
+
+@media (hover: none), (max-width: 1024px) {
+  :host {
+    --tcds-carousel-play-pause-opacity: 1;
+  }
+}
+
+:host(:hover),
+:host(:focus-within) {
+  --tcds-carousel-play-pause-opacity: 1;
+}
+
+[part="viewport"] {
+  grid-area: slides;
+  display: flex;
+  align-items: center;
+  overflow-x: scroll;
+  scroll-snap-type: x mandatory;
+  scrollbar-width: none;
+  overscroll-behavior: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  [part="viewport"] {
+    scroll-behavior: smooth;
+  }
+}
+
+[part="viewport"]::-webkit-scrollbar {
+  display: none;
+}
+
+::slotted(tcds-slide) {
+  flex: 1 0 100%;
+  scroll-snap-align: start;
+  scroll-snap-stop: always;
+}
+
+[part="navigation"] {
+  grid-area: navigation;
+  display: flex;
+  align-items: center;
+  justify-content: var(--tcds-carousel-navigation-justify, center);
+  gap: var(--tcds-space-loose);
+}
+
+[part="indicators"] {
+  display: flex;
+  gap: var(--tcds-space-x-loose);
+}
+
+[part="indicator"] {
+  display: inline-flex;
+  overflow: hidden;
+  height: 0.85rem;
+  width: 0.85rem;
+  border-radius: 0.85rem;
+  padding: 0;
+  border: 0;
+  cursor: pointer;
+  background-color: #a8a8a8;
+  transition: background-color var(--tcds-animation-productive-duration) var(--tcds-animation-productive-easing);
+}
+
+[part="indicator"]:hover,
+[part="indicator"][aria-expanded="true"] {
+  background-color: var(--tcds-color-primary);
+}
+
+[part="play-pause"] {
+  grid-area: play-pause;
+  opacity: var(--tcds-carousel-play-pause-opacity, 0);
+  transition: opacity var(--tcds-animation-productive-duration) var(--tcds-animation-productive-easing);
+}
+
+[part="next"]:not(:hover),
+[part="previous"]:not(:hover) {
+  --tcds-button-text-color: #a8a8a8;
+}
