+++ components/Icon/style.css
@@ -1,0 +1,29 @@
+/* stylelint-disable property-no-vendor-prefix */
+
+:host {
+  display: inline-flex;
+  vertical-align: sub;
+}
+
+[part="icon"] {
+  color: inherit;
+  height: var(--tcds-icon-size, 1em);
+  width: var(--tcds-icon-size, 1em);
+}
+
+[part="icon"]::before {
+  background: currentcolor;
+  content: "";
+  display: block;
+  height: 100%;
+  width: 100%;
+  -webkit-mask: var(--tcds-icon);
+  mask: var(--tcds-icon);
+  -webkit-mask-position: 50% 50%;
+  mask-position: 50% 50%;
+  -webkit-mask-repeat: no-repeat;
+  mask-repeat: no-repeat;
+  -webkit-mask-size: contain;
+  mask-size: contain;
+  vertical-align: middle;
+}
