+++ components/Tabs/style.css
@@ -1,0 +1,64 @@
+:host {
+  display: block;
+}
+
+:host([size*="small"]) {
+  --tcds-tabs-tab-font-size: var(--tcds-font-size-small);
+  --tcds-tabs-tab-height: var(--tcds-size-small);
+  --tcds-tabs-tab-x-padding: var(--tcds-space-tight);
+}
+
+:host([size*="large"]) {
+  --tcds-tabs-tab-flex-grow: 1;
+  --tcds-tabs-tab-font-size: var(--tcds-font-size-medium-plus);
+  --tcds-tabs-tab-height: var(--tcds-size-large);
+  --tcds-tabs-tab-x-padding: var(--tcds-space-loose);
+  --tcds-tab-padding-top: var(--tcds-space-loose);
+}
+
+[role="tablist"] {
+  display: flex;
+  overflow-x: auto;
+  overscroll-behavior: none;
+}
+
+[part="tab"] {
+  appearance: none;
+  background: none;
+  border: none;
+  color: var(--tcds-color-secondary);
+  font-family: var(--tcds-font-ui);
+  font-weight: var(--tcds-font-weight-semibold);
+  font-size: var(--tcds-tabs-tab-font-size, var(--tcds-font-size-medium));
+  flex: var(--tcds-tabs-tab-flex-grow, 0) 0;
+  padding: 0 var(--tcds-tabs-tab-x-padding, var(--tcds-space-normal));
+  height: var(--tcds-tabs-tab-height, var(--tcds-size-medium));
+  width: 100%;
+  white-space: nowrap;
+  cursor: pointer;
+  transition: color var(--tcds-animation-productive-duration) var(--tcds-animation-productive-easing);
+}
+
+[part="tab"]:nth-of-type(1) {
+  padding-left: 0;
+}
+
+[part="tab"] span {
+  border-bottom: 5px solid var(--tcds-tabs-tab-indicator-color, transparent);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  height: 100%;
+  transition: border-color var(--tcds-animation-productive-duration) var(--tcds-animation-productive-easing);
+}
+
+@media (any-hover: hover) {
+  [part="tab"]:hover {
+    color: var(--tcds-color-primary);
+  }
+}
+
+[part="tab"][aria-expanded="true"] {
+  --tcds-tabs-tab-indicator-color: var(--tcds-tabs-tab-indicator-active-color, var(--tcds-color-primary));
+}
