+++ components/Tabs/style.light.css
@@ -1,0 +1,15 @@
+tcds-tabs:not(:only-child) {
+  margin: var(--tcds-space-normal) 0;
+}
+
+tcds-tabs:first-child:not(:only-child) {
+  margin: 0 0 var(--tcds-space-normal);
+}
+
+.bg-primary[data-theme="dark"] tcds-tabs {
+  --tcds-tabs-tab-indicator-active-color: var(--tcds-color-tint);
+}
+
+[class*="bg-"][data-theme] tcds-tabs::part(tab) {
+  color: inherit;
+}
