+++ package.json
@@ -33,6 +33,8 @@
     "@babel/preset-env": "^7.19.1",
     "autoprefixer": "^10.4.11",
     "babel-loader": "^8.2.5",
+    "construct-style-sheets-polyfill": "^3.1.0",
+    "constructable-style-loader": "^1.0.0-alpha.8",
     "eslint": "^8.23.1",
     "gulp": "^4.0.2",
     "gulp-imagemin": "^8.0.0",
